// frontend/src/api/generated/model/lostFoundClaimStatus.ts
export type LostFoundClaimStatus = typeof LostFoundClaimStatus[keyof typeof LostFoundClaimStatus];


export const LostFoundClaimStatus = {
  new: 'new',
  accepted: 'accepted',
  declined: 'declined',
  referred: 'referred',
} as const;