// frontend/src/api/generated/model/lostFoundCustody.ts
export type LostFoundCustody = typeof LostFoundCustody[keyof typeof LostFoundCustody];


export const LostFoundCustody = {
  finder: 'finder',
  administration: 'administration',
} as const;