// frontend/src/api/generated/model/listLostFoundStatus.ts
export type ListLostFoundStatus = typeof ListLostFoundStatus[keyof typeof ListLostFoundStatus];


export const ListLostFoundStatus = {
  published: 'published',
  claimed: 'claimed',
} as const;